src/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
pyyaml>=6.0.0
python-dateutil>=2.8.2
tqdm>=4.65.0
matplotlib>=3.7.1
seaborn>=0.12.2
pytz>=2023.3
tabulate>=0.9.0
psutil>=5.9.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.11.1
chromadb>=0.4.18
psycopg2-binary>=2.9.5

# Embedding and NLP
sentence-transformers>=2.2.2
transformers>=4.35.2
scikit-learn>=1.3.0
spacy>=3.6.1
nltk>=3.8.1

# LLM clients
openai>=1.3.0
anthropic>=0.5.0
huggingface-hub>=0.17.3
tiktoken>=0.5.1

# Command line interface
click>=8.1.3
rich>=13.4.2
typer>=0.9.0

# API and web
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
httpx>=0.24.1

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
pylint>=2.17.4 